フッターのコンポーネントを切り出し、未読タブ上にのみフッターのリンクを表示するように実装
